// packages/engine/src/tables.generated.ts
export const FAMILIES = [
  "U",
  "D",
  "L",
  "R",
  "F",
  "B",
  "Uw",
  "Dw",
  "Lw",
  "Rw",
  "Fw",
  "Bw",
  "M",
  "E",
  "S",
  "x",
  "y",
  "z",
] as const;

export type Family = (typeof FAMILIES)[number];